--- src/BootstrapBlazor.Server/wwwroot/report/report.css
@@ -1,106 +1,0 @@
-
-
-.univer-fixed {
-    position: absolute;
-}
-
-ul, li {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-}
-.bb-formula-selection, .bb-edit-container {
-    font-size: 12px;
-}
-
-.bb-formula-selection > div {
-    display: flex;
-}
-    .bb-formula-selection .bb-category-list{
-        width: 50%;
-        margin-right: 10px;
-    }
-    .bb-formula-selection .bb-variable-list{
-        width: 50%;
-    }
-        .bb-formula-selection li, .bb-edit-container .bb-edit-tabs li {
-            height: 30px;
-            line-height: 30px;
-            padding-left: 8px;
-            font-size: 14px;
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-            cursor: pointer;
-            user-select: none;
-        }
-        .bb-formula-selection li:hover, .bb-edit-container .bb-edit-tabs li:hover {
-            background-color: #efefef;
-        }
-        .bb-formula-selection li.active, .bb-edit-container .bb-edit-tabs li.active {
-            background-color: #e3e3e3;
-        }
-        .bb-formula-selection li span {
-            font-size: 10px;
-            margin-left: 5px;
-            color: rgba(0, 0, 0, 0.6)
-        }
-    .bb-formula-selection input {
-        margin-top: 10px;
-        width: 80%;
-        height: 20px;
-        padding-left: 5px;
-    }
-    .bb-formula-selection button {
-        width: 60px;
-        height: 25px;
-        padding-left: 5px;
-        margin-left: 10px;
-        cursor: pointer;
-    }
-
-.bb-edit-container {
-    display: flex;
-}
-    .bb-edit-container .bb-edit-tabs {
-       width: 30%;
-    }
-    .bb-edit-container .bb-edit-content {
-       width: 70%;
-       padding-left: 5px;
-    }
-    .bb-edit-container .bb-edit-content h2 {
-       margin: 0;
-       font-size: 20px;
-       font-weight: normal;
-    }
-    .bb-edit-container .bb-edit-content h3 {
-       margin: 0;
-       font-size: 16px;
-       font-weight: normal;
-    }
-    .bb-edit-container .bb-edit-content label:has(+input[type="text"]) {
-       display: block;
-       font-size: 12px;
-    }
-    .bb-edit-container .bb-edit-content input[type="text"]:not(.variable):not(.valid-value),
-    .bb-edit-container .bb-edit-content select:not(.match-type):not(.action) {
-       width: 100%;
-       height: 25px;
-       box-sizing: border-box;
-    }
-
-    .bb-edit-container .bb-edit-content > div {
-        padding-left: 10px;
-    }
-
-    .bb-edit-container .bb-edit-content > div > div {
-        margin-bottom: 10px;
-    }
-    .bb-edit-container .bb-edit-content .bb-valid-wrapper li {
-        display: flex;
-    }
-    .bb-edit-container .bb-edit-content .bb-valid-wrapper li input[type="text"], .bb-edit-container .bb-edit-content .bb-valid-wrapper li select {
-        flex: 1;
-        margin-right: 3px;
-    }
